Made links highlight

+++ app/javascript/react/containers/ArmiesIndexContainer.js
@@ -36,34 +36,36 @@ class ArmiesIndexContainer extends Component {
 	}
 
 	deleteFromDatabase() {
-		fetch(`/api/v1/armies/${this.state.idToDelete}`, {
-			method: 'DELETE',
-			credentials: 'same-origin',
-        	headers: {'Content-Type': 'application/json'}
-		})
-		.then(response => {
-			if (response.ok) {
-				return response
-			} else {
-				let errorMessage = `${response.status} (${response.statusText})`,
-				error = new Error(errorMessage)
-				throw(error)
-			}
-		})
-		.then(response => response.json())
-		.then(body => {
-			let updatedArmies = []
-			let i
-			for (i = 0; i < this.state.armies.length; i++) {
-				if (this.state.armies[i].id !== body.army.id) {
-					updatedArmies.push(this.state.armies[i])
+		if (window.confirm('Delete?')) {
+			fetch(`/api/v1/armies/${this.state.idToDelete}`, {
+				method: 'DELETE',
+				credentials: 'same-origin',
+	        	headers: {'Content-Type': 'application/json'}
+			})
+			.then(response => {
+				if (response.ok) {
+					return response
+				} else {
+					let errorMessage = `${response.status} (${response.statusText})`,
+					error = new Error(errorMessage)
+					throw(error)
 				}
-			}
-			this.setState({ armies: updatedArmies })
-		})
-		.catch(error => console.error(`Error in fetch: ${error.message}`))
-		this.hideDeletionTile()
-		this.showDeletionSuccessMessage()
+			})
+			.then(response => response.json())
+			.then(body => {
+				let updatedArmies = []
+				let i
+				for (i = 0; i < this.state.armies.length; i++) {
+					if (this.state.armies[i].id !== body.army.id) {
+						updatedArmies.push(this.state.armies[i])
+					}
+				}
+				this.setState({ armies: updatedArmies })
+			})
+			.catch(error => console.error(`Error in fetch: ${error.message}`))
+			this.hideDeletionTile()
+			this.showDeletionSuccessMessage()
+		}
 	}	
 
 	showDeletionTile(id, name) {
